\name{mmtfa-package}
\alias{mmtfa-package}
\alias{mmtfapackage}
\alias{MMtFA}
\docType{package}
\title{
mmtfa: Model-Based Clustering and Classification with Mixtures of Modified Multivariate t Factor Analyzers
}
\description{
Fits mixtures of multivariate modified t-factor analyzers via the alternating expectation-conditional maximization algorithm under a clustering or classification paradigm in serial or parallel.
}
\details{
\tabular{ll}{
Package: \tab mmtfa\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2018-06-11\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey L. Andrews, Paul D. McNicholas, and Mathieu Chalifour

Maintained by: Jeffrey L. Andrews <jeff.andrews@ubc.ca>
}
\references{
Andrews JL and McNicholas PD (2011a), 'Extending mixtures of multivariate t-factor analyzers'.
\emph{Statistics and Computing} 21(3), 361--373.

Andrews JL and McNicholas PD (2011b), 'Mixtures of modified t-factor analyzers for model-based
clustering, classification, and discriminant analysis'. \emph{Journal of Statistical Planning and Inference} 141(4), 1479--1486.
}
\keyword{ package }
\seealso{
\code{\link{mmtfa}} for main function
}
